andando con algunas excepciones

+++ ComiteAccesoADatos/EF/RepositorioUsuario.cs
@@ -1,4 +1,5 @@
 ﻿using ComiteLogicaNegocio.Entidades;
+using ComiteLogicaNegocio.Excepciones.Usuario;
 using ComiteLogicaNegocio.InterfacesRepositorios;
 using Microsoft.EntityFrameworkCore.SqlServer.Query.Internal;
 
@@ -17,7 +18,11 @@ namespace ComiteAccesoADatos.EF
         {
             if (obj == null) 
             {
-                throw new ArgumentNullException("No se recibio el usuario vàlido");
+                throw new UsuarioException("No se recibio el usuario vàlido");
+            }
+            if (UsuarioExiste(obj.Email.Value))
+            {
+                throw new UsuarioException("El usuario ya existe");
             }
             _context.Add(obj);
             _context.SaveChanges();
@@ -36,7 +41,7 @@ namespace ComiteAccesoADatos.EF
                 .FirstOrDefault(usuario => usuario.ID == id);
             if (unUsuario == null)
             {
-                throw new Exception($"No se encontro la el usuario con id {id}");
+                throw new UsuarioException($"No se encontro la el usuario con id {id}");
             }
             return unUsuario;
         }
@@ -50,25 +55,63 @@ namespace ComiteAccesoADatos.EF
                 .FirstOrDefault(usuario => usuario.Email.Value == email);
             if (unUsuario == null)
             {
-                throw new Exception($"No se encontro la el usuario con email {email}");
+                throw new UsuarioException($"No se encontro la el usuario con email {email}");
             }
             return unUsuario;
         }
+        public bool UsuarioExiste(string email)
+        {
 
+            Usuario? unUsuario = null;
+            unUsuario =
+                _context.usuarios
+                .AsEnumerable()
+                .FirstOrDefault(usuario => usuario.Email.Value == email);
+            return unUsuario != null;
+        }
         public void Delete(Usuario obj)
         {
-            Usuario u = GetByEmail(obj.Email.Value);
-            _context.usuarios.Remove(u);
-            _context.SaveChanges();
+            try
+            {
+                Usuario u = GetByEmail(obj.Email.Value);
+                if (u == null)
+                {
+                    throw new UsuarioException("No se recibio el usuario vàlido");
+                }
+                _context.usuarios.Remove(u);
+                _context.SaveChanges();
+            }
+            catch (Exception)
+            {
+
+                throw new UsuarioException("Error al eliminar usuario");
+            }
+
         }
 
         public void Edit(Usuario obj)
         {
             Usuario u = GetById(obj.ID);
-            u.Email = obj.Email;
-            u.Password = obj.Password;
-            _context.usuarios.Update(u);
-            _context.SaveChanges();
+            if (u == null)
+            {
+                throw new UsuarioException("No se recibio el usuario vàlido");
+            }
+            if (UsuarioExiste(obj.Email.Value)) {
+                throw new UsuarioException("El usuario ya existe");
+            }
+            try
+            {
+                u.Email = obj.Email;
+                u.Password = obj.Password;
+                _context.usuarios.Update(u);
+                _context.SaveChanges();
+            }
+            catch (Exception)
+            {
+
+                throw new UsuarioException("Error al editar usuario");
+            }
+
         }
     }
 }
